+++ malachite-nz/src/bin_util/demo_and_bench/natural_polynomial/arithmetic/mod.rs
@@ -16,6 +16,7 @@ pub(crate) fn register(runner: &mut Runner) {
     is_unit::register(runner);
     make_monic::register(runner);
     mod_is_reduced::register(runner);
+    mod_neg::register(runner);
     mod_op::register(runner);
     mod_power_of_2::register(runner);
     mod_power_of_2_neg::register(runner);
@@ -28,6 +29,7 @@ mod height;
 mod is_unit;
 mod make_monic;
 mod mod_is_reduced;
+mod mod_neg;
 mod mod_op;
 mod mod_power_of_2;
 mod mod_power_of_2_neg;